IIIIIIlIIIIII

+++ src/main/java/icu/suc/megawalls78/event/EnergyChangeEvent.java
@@ -8,30 +8,30 @@ import org.jetbrains.annotations.NotNull;
 
 public class EnergyChangeEvent extends PlayerEvent implements Cancellable {
 
-    private final int origin;
-    private final int energy;
-    private final int max;
+    private final float origin;
+    private final float energy;
+    private final float max;
 
     private boolean cancelled;
 
     private static final HandlerList handlers = new HandlerList();
 
-    public EnergyChangeEvent(Player player, int origin, int energy, int max) {
+    public EnergyChangeEvent(Player player, float origin, float energy, float max) {
         super(player);
         this.origin = origin;
         this.energy = energy;
         this.max = max;
     }
 
-    public int getOrigin() {
+    public float getOrigin() {
         return origin;
     }
 
-    public int getEnergy() {
+    public float getEnergy() {
         return energy;
     }
 
-    public int getMax() {
+    public float getMax() {
         return max;
     }
